refactor: 모든 형식 검사 RegisterModel로 이동
