+++ FlagBoard/FlagBoard/Controller/Login/EmailVerifyController.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 import Alamofire
-import SwiftyJSON
 
 class EmailVerifyController: UIViewController {
