+++ FlagBoard/FlagBoard/Controller/Login/RegisterEmailAndInformationController.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 import Alamofire
-import SwiftyJSON
 
 
 class RegisterEmailAndInformationController: UIViewController {
@@ -122,7 +121,7 @@ class RegisterEmailAndInformationController: UIViewController {
         if userName.isEmpty {
             print("이름이 비었습니다!")
             return false
-        } else if !isValidName(name: userName) {
+        } else if !RegisterModel.isValidName(name: userName) {
             print("이름의 형식이 틀렸습니다!")
             return false
         }
@@ -136,7 +135,7 @@ class RegisterEmailAndInformationController: UIViewController {
         if userStudentId.isEmpty {
             print("학번이 비었습니다!")
             return false
-        } else if !isValidStudentId(studentId: userStudentId) {
+        } else if !RegisterModel.isValidStudentId(studentId: userStudentId) {
             print("학번의 형식이 틀렸습니다!")
             return false
         }
@@ -151,37 +150,12 @@ class RegisterEmailAndInformationController: UIViewController {
         if userEmail.isEmpty {
             print("이메일이 비었습니다!")
             return false
-        } else if !isValidEmail(email: userSuwonEmail) {
+        } else if !RegisterModel.isValidEmail(email: userSuwonEmail) {
             print("이메일의 형식이 틀렸습니다!")
             return false
         }
         
         return true
-    }
-    
-    
-    // 이메일 형식 검사
-    func isValidEmail(email: String?) -> Bool {
-        let emailRegEx = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}"
-        
-        let userEmail = NSPredicate(format:"SELF MATCHES %@", emailRegEx)
-                  return userEmail.evaluate(with: email)
-    }
-    
-    // 이름 형식 검사
-    func isValidName(name: String?) -> Bool {
-        let nameRegEx = "[가-힣]{2,5}"
-        
-        let userName = NSPredicate(format:"SELF MATCHES %@", nameRegEx)
-                  return userName.evaluate(with: name)
-    }
-    
-    // 학번 형식 검사
-    func isValidStudentId(studentId: String?) -> Bool {
-        let studentIdRegEx = "[0-9]{8}"
-        
-        let userStudentId = NSPredicate(format:"SELF MATCHES %@", studentIdRegEx)
-                  return userStudentId.evaluate(with: studentId)
     }
     
     func pushToNextVC(email: String) {
